Style in docker scripts
